Make plot_cross_validation a component of CrossValidationItem
