+++ skore/src/skore/cross_validate.py
@@ -6,42 +6,9 @@ function in order to enrich it with more information and enable more analysis.
 
 from typing import Literal
 
+from skore import logger
 from skore.item.cross_validate_item import CrossValidationItem
 from skore.project import Project
-
-
-def plot_cross_validation(cv_results: dict):
-    """Plot the result of a cross-validation run."""
-    import altair
-    import pandas
-
-    # df = pandas.DataFrame(cv_results).melt(var_name="metric", value_name="score")
-    df = (
-        pandas.DataFrame(cv_results)
-        .reset_index(names="split")
-        .melt(id_vars="split", var_name="metric", value_name="score")
-    )
-
-    input_dropdown = altair.binding_select(
-        options=df["metric"].unique().tolist(), name="Metric: "
-    )
-    selection = altair.selection_point(
-        fields=["metric"], bind=input_dropdown, value="test_score"
-    )
-
-    return (
-        altair.Chart(df, title="Cross-validation scores per split")
-        .mark_bar()
-        .encode(
-            altair.X("split:N").title("Split number"),
-            altair.Y("score:Q").title("Score"),
-            tooltip=["metric:N", "split:N", "score:Q"],
-        )
-        .interactive()
-        .add_params(selection)
-        .transform_filter(selection)
-        .properties(width=400, height=200)
-    )
 
 
 def _find_ml_task(
@@ -207,7 +174,10 @@ def cross_validate(
     if project is not None:
         project.put_item("cross_validation", cross_validation_item)
 
-    plot_cross_validation(cross_validation_item.cv_results)
+    try:
+        display(cross_validation_item.plot)
+    except NameError:
+        logger.warn("Displaying the plot failed because 'display' is not in scope.")
 
     # Remove information related to our scorers, so that our return value is
     # the same as sklearn's
